lancelot: pull get_operand_target into helpers
