+++ capa/features/extractors/lancelot/insn.py
@@ -26,7 +26,9 @@ from lancelot import (
 import capa.features.extractors.helpers
 from capa.features import ARCH_X32, ARCH_X64, MAX_BYTES_FEATURE_SIZE, Bytes, String, Characteristic
 from capa.features.insn import Number, Offset, Mnemonic
+from capa.features.extractors.lancelot.helpers import get_operand_target
 from capa.features.extractors.lancelot.function import get_call_graph
+from capa.features.extractors.lancelot.indirect_calls import NotFoundError, resolve_indirect_call
 
 logger = logging.getLogger(__name__)
 
@@ -64,33 +66,6 @@ def get_imports(xtor):
                 impname = imp.name.decode("ascii")
                 imports[imp.address] = "%s.%s" % (libname, impname)
     return imports
-
-
-def get_operand_target(insn, op):
-    if insn.address in (0x455A41, 0x46AE12):
-        print(insn, insn.operands)
-
-    if op[OPERAND_TYPE] == OPERAND_TYPE_MEMORY:
-        # call direct, x64
-        # rip relative
-        # kernel32-64:180001041    call    cs:__imp_RtlVirtualUnwind_0
-        if op[MEMORY_OPERAND_BASE] == "rip":
-            return op[MEMORY_OPERAND_DISP] + insn.address + insn.length
-
-        # call direct, x32
-        # mimikatz:0x403BD3  call    ds:CryptAcquireContextW
-        elif op[MEMORY_OPERAND_BASE] == None:
-            return op[MEMORY_OPERAND_DISP]
-
-    # call via thunk
-    # mimikatz:0x455A41  call    LsaQueryInformationPolicy
-    elif op[OPERAND_TYPE] == OPERAND_TYPE_IMMEDIATE and op[IMMEDIATE_OPERAND_IS_RELATIVE]:
-        return op[IMMEDIATE_OPERAND_VALUE] + insn.address + insn.length
-
-    elif op[OPERAND_TYPE] == OPERAND_TYPE_IMMEDIATE:
-        return op[IMMEDIATE_OPERAND_VALUE]
-
-    raise ValueError("memory operand has no target")
 
 
 @lru_cache
